Fixed crash on resize, started cpp GUI work

+++ src/camera.rs
@@ -101,9 +101,9 @@ impl CameraController {
         }
     }
 
-    pub fn process_mouse(&mut self, mouse_dx: f64, mouse_dy: f64) {
-        self.rotate_horizontal = mouse_dx as f32;
-        self.rotate_vertical = mouse_dy as f32;
+    pub fn process_mouse(&mut self, mouse_dx: f32, mouse_dy: f32) {
+        self.rotate_horizontal = mouse_dx;
+        self.rotate_vertical = mouse_dy;
     }
 
     pub fn update_camera(&mut self, camera: &mut Camera, dt: std::time::Duration) {
